Clean & Refactoring data-types in Worklists
