+++ src/app/worklists/services/activity.service.ts
@@ -11,86 +11,25 @@ import { Observable } from 'rxjs/Observable';
 
 import { CoreService } from '../../core';
 
-import { Task } from '../data-types/task';
-import { Activity, ProjectRef, ResourceRef, PersonRef,TaskRef } from '../data-types/project';
-import { ProcessModel } from '../data-types/project';
+import { ClosedTask, Task, TaskRef } from '../data-types/task';
+import { Activity, ProjectRef, ResourceRef } from '../data-types/project';
 
 @Injectable()
 export class ActivityService {
 
   public constructor(private core: CoreService) { }
-
-  public getResourcesList(projectUID: string): Observable<ResourceRef[]> {
-    const path = `v1/project-management/projects/${projectUID}/resources`;
-
-    return this.core.http.get<ResourceRef[]>(path);
-  }
-
-  public getRequestersList(projectUID: string): Observable<PersonRef[]> {
-    const path ='v1/project-management/projects/'+ projectUID +'/requesters';
-
-    return this.core.http.get<PersonRef[]>(path);
-  }
-
-   public getResponsiblesList(projectUID: string): Observable<PersonRef[]> {
-    const path ='v1/project-management/projects/'+ projectUID +'/responsibles';
-
-    return this.core.http.get<PersonRef[]>(path);
-  }
-
-  public getTaskManagers(projectUID: string): Observable<PersonRef[]> {
-    const path = 'v1/project-management/projects/'+ projectUID + '/task-managers';
-
-    return this.core.http.get<PersonRef[]>(path);
-  }   
-
-  public getActivity(itemId: number): Observable<any> {    
-    const path = `v1/project-management/activities/${itemId}`;
-
-    return this.core.http.get<any>(path);
-  }
-
-  public getTasks(itemId: number): Observable<TaskRef[]> {
-    const path =  `v1/project-management/activities/${itemId}/tasks`;
-
-    return this.core.http.get<TaskRef[]>(path);
-
-  }
-
-  public getProcess(): Observable<ProcessModel[]> {
-    const path ='v1/projects/process-models/for-activities';
-
-    return this.core.http.get<ProcessModel[]>(path);
-  }
-
-  public getEvents(): Observable<ProcessModel[]> {
-    const path ='v1/projects/process-models/for-events';
-    
-    return this.core.http.get<ProcessModel[]>(path);
-
-  }
-
-  public addProcessModel(projectUID: string, processModelUID: string, 
-                         activity: Activity): Observable<Activity> {
-
-    const path = `v1/project-management/projects/${projectUID}/create-from-process-model/${processModelUID}`;
-
-    return this.core.http.post<Activity>(path, activity);
-  }
-
+ 
   public addManualActivity(projectUID:string, activity:Activity): Observable<any[]> {
     const path = `v1/project-management/projects/${projectUID}/activities`;
 
     return this.core.http.post<any[]>(path, activity);
   }
 
-  public updateActivity(projectUID:string, activityId: number, activity: Activity): 
-                        Observable<Activity> {
-    const path = `v1/project-management/projects/${projectUID}/activities/${activityId}`;
-    
-    return this.core.http.put<Activity>(path,activity);
-    
-  }
+  public getActivity(itemId: number): Observable<any> {    
+    const path = `v1/project-management/activities/${itemId}`;
+
+    return this.core.http.get<any>(path);
+  }  
 
   public searchActivities(projectUID: string, filter: object,
                           orderBy: string, keywords: string): Observable<Task[]> {
@@ -100,5 +39,36 @@ export class ActivityService {
 
     return this.core.http.get<Task[]>(path);                      
   }
+
+  public getActivitiesListAsGantt(projectId: string): Promise<Task[]> {
+    const path = `v1/project-management/projects/${projectId}/activities/as-gantt`;    
+
+    return this.core.http.get<Task[]>(path)
+                         .toPromise();
+
+  }
+
+  public getActivities(projectId: string): Promise<Task[]> {
+    const path = `v1/project-management/projects/${projectId}/activities`;
+                 
+    return this.core.http.get<Task[]>(path)
+                         .toPromise();
+
+  }
+
+  public updateActivity(projectUID:string, activityUID: string, task: TaskRef): 
+  Observable<TaskRef> {
+  const path = `v1/project-management/projects/${projectUID}/activities/${activityUID}`;
+
+
+  return this.core.http.put<TaskRef>(path,task);
+}
+
+public closeActivity(projectUID:string, activityUID:string, closeTask: ClosedTask): Observable<any[]> {
+  const path = `v1/project-management/projects/${projectUID}/activities/${activityUID}/close`;
+                
+  return this.core.http.post<any[]>(path, closeTask);
+}
+
 }
